AwsAuthService/requirements.txt
aws-cdk.core
aws-cdk.aws_cognito
aws-cdk.aws_lambda
boto3
botocore